requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
python-multipart
liac-arff
